gobjwork: CalcStatus drop redundant short cast on equipped item (94.62%)

+++ src/gobjwork.cpp
@@ -1900,7 +1900,7 @@ void CCaravanWork::CalcStatus()
 	for (int equipIdx = 0; equipIdx < 4; equipIdx++) {
 		int equipSlot = m_equipment[equipIdx];
 		if (equipSlot >= 0) {
-			int itemIdx = (short)m_inventoryItems[equipSlot];
+			int itemIdx = m_inventoryItems[equipSlot];
 			int itemType = GetItemDataPtr(itemIdx)[0];
 
 			if (itemType == 1) {
